Fix deadlock in audio termination
Calling subscribe or unsubscribe while the callback is being called
is already protected on naoqi side. So no need to protect it on the bridge
side, this is what previously led to a deadlock.
We only need mutex protection on configuration variable (publishing,
recording, logging) and also make sure calling subscribe and unsubscribe
at the same time is not possible (even though this is also protected in
naoqi).

Change-Id: Iae604c047046fec9e24832dd4df5017ff4ae724f